+++ pokedex/src/routes/Routers.js
@@ -1,8 +1,9 @@
 import react from "react";
 import { Route,Routes, BrowserRouter } from "react-router-dom";
-import { Home } from "../pages/Home";
-import { Pokedex } from "../pages/Pokedex";
-import { PokemonDetail } from "../pages/PokemonDetail";
+import { Home } from "../pages/Home/Home";
+import { Pokedex } from "../pages/Pokedex/Pokedex";
+import { PokemonDetail } from "../pages/PokemonDetail/PokemonDetail";
+
 
 export const Routers = () => {
     return(
